requirements.txt

# Google Genai - Nueva API directa (para llm_service_direct.py)
google-genai

# Database
pymongo==4.6.0

# HTTP requests and web scraping
requests==2.31.0
beautifulsoup4==4.12.0

# Date utilities
python-dateutil==2.8.2

# JSON handling (built-in, but explicit for clarity)
# json - built-in module

# Logging and error handling (built-in modules)
# logging - built-in module
# traceback - built-in module

# Environment variables
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.4
pytest-mock==3.12.0
pytest-asyncio==0.23.2
mongomock==4.1.2